@@ -3,58 +3,58 @@ description: Ecommerce App
 
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+1
+version: 1.1.0
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
-  cached_network_image: ^3.2.3
-  cached_network_svg_image: ^0.0.7
-  change_app_package_name: ^1.1.0
-  cloud_firestore: ^4.9.0
-  cupertino_icons: ^1.0.5
+  cached_network_image: ^3.4.1
+  cached_network_svg_image: ^1.1.0
+  change_app_package_name: ^1.4.0
+  cloud_firestore: ^5.5.1
+  cupertino_icons: ^1.0.8
   dartz: ^0.10.1
-  dropdown_search: ^5.0.6
-  firebase_auth: ^4.8.0
-  firebase_core: ^2.15.0
-  firebase_crashlytics: ^3.3.5
-  firebase_messaging: ^14.6.5
+  dropdown_search: ^6.0.1
+  firebase_auth: ^5.3.4
+  firebase_core: ^3.8.1
+  firebase_crashlytics: ^4.2.0
+  firebase_messaging: ^15.1.6
   flutter:
     sdk: flutter
-  flutter_local_notifications: ^15.1.0+1
-  flutter_map: ^5.0.0
-  flutter_native_splash: ^2.3.2
-  flutter_otp_text_field: ^1.1.1
+  flutter_local_notifications: ^18.0.1
+  flutter_map: ^7.0.2
+  flutter_native_splash: ^2.4.3
+  flutter_otp_text_field: ^1.2.0
   flutter_paypal_checkout: ^1.0.1
-  flutter_ringtone_player: ^3.2.0
-  flutter_screenutil: ^5.9.0
-  flutter_stripe: ^9.3.0
-  flutter_svg: ^2.0.7
-  geocoding: ^2.1.0
-  geolocator: ^9.0.2
-  get: ^4.6.5
+  flutter_ringtone_player: ^4.0.0
+  flutter_screenutil: ^5.9.3
+  flutter_stripe: ^11.3.0
+  flutter_svg: ^2.0.16
+  geocoding: ^3.0.0
+  geolocator: ^13.0.2
+  get: ^4.6.6
   get_storage: ^2.1.1
   google_polyline_algorithm: ^3.1.0
-  icons_launcher: ^2.1.3
-  jiffy: ^6.2.1
-  latlong2: ^0.9.0
-  lottie: ^2.5.0
+  icons_launcher: ^3.0.0
+  jiffy: ^6.3.2
+  latlong2: ^0.9.1
+  lottie: ^3.3.0
   otp_text_field: ^1.1.3
-  path: ^1.8.3
-  pay: ^1.1.2
+  path: ^1.9.0
+  pay: ^2.0.0
   rating_dialog: ^2.0.4
-  responsive_builder: ^0.7.0
-  routing_client_dart: ^0.4.3
+  responsive_builder: ^0.7.1
+  routing_client_dart: ^0.5.5
   # sentry_flutter: ^7.9.0
-  shared_preferences: ^2.2.0
-  url_launcher: ^6.1.12
+  shared_preferences: ^2.3.3
+  url_launcher: ^6.3.1
 
 dependency_overrides:
   http: ^1.0.0
 
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
 # flutter pub run change_app_package_name:main com.amadytech.happyshop
