@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class AddAddressEdit extends StatelessWidget {
-  const AddAddressEdit({Key? key}) : super(key: key);
+  const AddAddressEdit({super.key});
 
   @override
   Widget build(BuildContext context) {
