@@ -8,7 +8,7 @@ import 'package:get/get.dart';
 import 'package:happyshop/controller/cart/paymentpagecontroller.dart';
 
 class PaymentPage extends StatelessWidget {
-  const PaymentPage({Key? key}) : super(key: key);
+  const PaymentPage({super.key});
   @override
   Widget build(BuildContext context) {
     PaymentPageController controller = Get.put(PaymentPageController());
