@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class SuccessResetpassword extends StatelessWidget {
-  const SuccessResetpassword({Key? key}) : super(key: key);
+  const SuccessResetpassword({super.key});
 
   @override
   Widget build(BuildContext context) {
