@@ -9,7 +9,7 @@ import 'package:get/get.dart';
 import 'package:jiffy/jiffy.dart';
 
 class Notifications extends StatelessWidget {
-  const Notifications({Key? key}) : super(key: key);
+  const Notifications({super.key});
 
   @override
   Widget build(BuildContext context) {
