@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class CustomTitleAuth extends StatelessWidget {
   final String title;
-  const CustomTitleAuth({Key? key, required this.title}) : super(key: key);
+  const CustomTitleAuth({super.key, required this.title});
 
   @override
   Widget build(BuildContext context) {
