@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class CustomDotControllerOnBoarding extends StatelessWidget {
-  const CustomDotControllerOnBoarding({Key? key}) : super(key: key);
+  const CustomDotControllerOnBoarding({super.key});
 
   @override
   Widget build(BuildContext context) {
