@@ -1,8 +1,8 @@
-import 'package:happyshop/controller/orders/order_details_controller.dart';
-import 'package:happyshop/core/constant/color.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:get/get.dart';
+import 'package:happyshop/controller/orders/order_details_controller.dart';
+import 'package:happyshop/core/constant/color.dart';
 
 class CustomOrdersDetailsMap extends StatelessWidget {
   const CustomOrdersDetailsMap({
@@ -18,9 +18,9 @@ class CustomOrdersDetailsMap extends StatelessWidget {
           onTap: (tapPosition, point) {
             controller.getPositionOfOrder();
           },
-          center: controller.latLng!,
-          zoom: 14.5),
-      nonRotatedChildren: const [],
+          initialCenter: controller.latLng!,
+          initialZoom: 14.5),
+      // nonRotatedChildren: const [],
       children: [
         TileLayer(
           urlTemplate: 'https://tile.openstreetmap.org/{z}/{x}/{y}.png',
@@ -31,11 +31,11 @@ class CustomOrdersDetailsMap extends StatelessWidget {
               point: controller.latLng!,
               width: 80,
               height: 80,
-              builder: (context) => const Icon(
-                    Icons.location_on,
-                    size: 40,
-                    color: AppColor.primaryColor,
-                  ))
+              child: Icon(
+                Icons.location_on,
+                size: 40,
+                color: AppColor.primaryColor,
+              ))
         ]),
       ],
     );
